+++ js/show-window.js
@@ -1,9 +1,9 @@
 const mainElement = document.querySelector(`.central`);
 
-const showWindow = (showElement) => {
+const showWindow = (view) => {
   const mainChildrens = mainElement.children;
   Array.from(mainChildrens).forEach((element) => element.remove());
-  mainElement.appendChild(showElement);
+  mainElement.appendChild(view.element);
 };
 
 export default showWindow;
