--- js/templates/rules.js
@@ -1,63 +1,0 @@
-import showWindow from '../show-window.js';
-import getElementFromTemplate from '../utils.js';
-import {game1Template, showGame2Template} from './game-1.js';
-import {introTemplate, showGreetingTemplate} from './intro.js';
-import {headerBackTemplate} from './header.js';
-
-const screenTemplate = (data) => `<header class="header">
-${headerBackTemplate}
-</header>
-<div class="rules">
-<h1 class="rules__title">Правила</h1>
-<p class="rules__description">Угадай ${data.numberLevels} раз для каждого изображения фото <img
-  src="img/photo_icon.png" width="16" height="16"> или рисунок <img
-  src="img/paint_icon.png" width="16" height="16" alt="">.<br>
-  Фотографиями или рисунками могут быть оба изображения.<br>
-  На каждую попытку отводится ${data.time} секунд.<br>
-  Ошибиться можно не более  ${data.lives} раз.<br>
-  <br>
-  Готовы?
-</p>
-<form class="rules__form">
-  <input class="rules__input" type="text" placeholder="Ваше Имя">
-  <button class="rules__button  continue" type="submit" disabled>Go!</button>
-</form>
-</div>
-<footer class="footer">
-<a href="https://htmlacademy.ru" class="social-link social-link--academy">HTML Academy</a>
-<span class="footer__made-in">Сделано в <a href="https://htmlacademy.ru" class="footer__link">HTML Academy</a> &copy; 2016</span>
-<div class="footer__social-links">
-  <a href="https://twitter.com/htmlacademy_ru" class="social-link  social-link--tw">Твиттер</a>
-  <a href="https://www.instagram.com/htmlacademy/" class="social-link  social-link--ins">Инстаграм</a>
-  <a href="https://www.facebook.com/htmlacademy" class="social-link  social-link--fb">Фэйсбук</a>
-  <a href="https://vk.com/htmlacademy" class="social-link  social-link--vk">Вконтакте</a>
-</div>
-</footer>`;
-
-function rulesTemplate(game) {
-  return getElementFromTemplate(screenTemplate(game));
-}
-
-const showGame1Template = (game) => {
-  const controlElementRules = document.querySelector(`.rules__button`);
-  const inputGame1 = document.querySelector(`.rules__input`);
-  const backButton = document.querySelector(`.back`);
-
-  backButton.addEventListener(`click`, () => {
-    showWindow(introTemplate);
-    showGreetingTemplate();
-  });
-
-  inputGame1.addEventListener(`input`, () => {
-    if (inputGame1.value) {
-      controlElementRules.removeAttribute(`disabled`);
-    }
-  });
-
-  controlElementRules.addEventListener(`click`, () => {
-    showWindow(game1Template(game));
-    showGame2Template(game);
-  });
-};
-
-export {rulesTemplate, showGame1Template};
