+++ js/view.js
@@ -1,0 +1,25 @@
+import getElementFromTemplate from './utils.js';
+
+class AbstractView {
+  get template() {
+
+  }
+
+  render() {
+    return getElementFromTemplate(this.template);
+  }
+
+  bind() {
+
+  }
+
+  get element() {
+    if (!this._element) {
+      this._element = this.render();
+      this.bind();
+    }
+    return this._element;
+  }
+}
+
+export default AbstractView;
